Move leader/follower utilities from server to helpers package
