+++ skipruntime-ts/helpers/src/index.ts
@@ -9,6 +9,6 @@ export {
   PolledExternalService,
   type PolledHTTPResource,
 } from "./external.js";
-export { SkipExternalService } from "./remote.js";
+export { SkipExternalService, asLeader, asFollower } from "./remote.js";
 export { SkipServiceBroker, fetchJSON, type Entrypoint } from "./rest.js";
 export { Count, Max, Min, Sum } from "./utils.js";
