+++ skipruntime-ts/server/package.json
@@ -20,7 +20,6 @@
   },
   "dependencies": {
     "@skipruntime/core": "0.0.13",
-    "@skipruntime/helpers": "0.0.13",
     "express": "^4.21.1"
   },
   "optionalDependencies": {
